Attempted to make Django project functional and link views
This commit reflects an unsuccessful attempt to set up and make the Django project functional. Explored various strategies for linking views, but encountered challenges preventing successful implementation.

This commit is a work in progress and should not be considered final. Further work and troubleshooting are required to address the underlying issues and achieve the desired functionality.